sdis/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector store and embeddings
faiss-cpu==1.7.4
openai==1.6.1
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Document processing
PyMuPDF==1.23.14
python-docx==1.1.0
python-multipart==0.0.6

# NLP and PII detection
spacy==3.7.2

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# File handling and utilities
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0

# Optional: Redis for caching
redis==5.0.1

# Optional: AWS SDK
boto3==1.34.34

# Logging and monitoring
structlog==23.2.0

# Background task processing
celery==5.3.4

# Environment variable loading
python-dotenv==1.0.0